rename package to avoid camel case in package names
