+++ integration/spark/src/main/spark3/java/io/openlineage/spark3/agent/lifecycle/plan/column/OutputFieldsCollector.java
@@ -1,4 +1,4 @@
-package io.openlineage.spark3.agent.lifecycle.plan.columnLineage;
+package io.openlineage.spark3.agent.lifecycle.plan.column;
 
 import io.openlineage.spark.agent.util.ScalaConversionUtils;
 import java.util.List;
