+++ integration/spark/src/test/spark3/java/io/openlineage/spark3/agent/lifecycle/plan/column/ColumnLevelLineageUtilsNonV2CatalogTest.java
@@ -1,4 +1,4 @@
-package io.openlineage.spark3.agent.lifecycle.plan.columnLineage;
+package io.openlineage.spark3.agent.lifecycle.plan.column;
 
 import static org.junit.jupiter.api.Assertions.assertTrue;
 import static org.mockito.Mockito.mock;
