+++ integration/spark/src/test/spark3/java/io/openlineage/spark3/agent/lifecycle/plan/column/ExpressionDependencyCollectorTest.java
@@ -1,4 +1,4 @@
-package io.openlineage.spark3.agent.lifecycle.plan.columnLineage;
+package io.openlineage.spark3.agent.lifecycle.plan.column;
 
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.times;
